add exec.sh

+++ deploy/exec.sh
@@ -8,7 +8,9 @@ FLG=$3
 
 LOGFILE=app.log
 
-rm ${LOGFILE}
+if [ -e LOGFILE ]; then
+  rm ${LOGFILE}
+fi
 
 # バイナリ実行
 ./code-pipeline-example -name="${NAME}" -num="${NUM}" -flg="${FLG}" > ${LOGFILE} 2>&1
